refactor: fixed warnings under -Wconversion and clang-tidy

+++ branch/hashed_perceptron/hashed_perceptron.cc
@@ -62,7 +62,7 @@ bool hashed_perceptron::predict_branch(uint64_t pc)
                    auto x = pc ^ (ghist_words[most_words] & champsim::msl::bitmask(last_word));
 
                    // XOR up to the next-to-the-last word
-                   x = std::accumulate(std::begin(ghist_words), std::next(std::begin(ghist_words), most_words), x, std::bit_xor<>{});
+                   x = std::accumulate(std::begin(ghist_words), std::next(std::begin(ghist_words), (long)most_words), x, std::bit_xor<>{});
 
                    return x & champsim::msl::bitmask(champsim::msl::lg2(TABLE_SIZE)); // stay within the table size
                  });
